Add support for Python 3.6 and move to Poetry
The project relies on the dataclasses package. This package is part of
the stdlib as of Python 3.7.

In order to support Python 3.6 this package must be installed from Pypi.
This backport on Pypi doesn't work with Python 3.7. To quote the README
from the package:

	"As of version 0.4, this code no longer works with Python 3.7.
	For 3.7, use the built-in dataclasses module."

The traditional requirements file doesn't allow to install certain
packages based on the Python version and don't install the package for
other Python versions.

Poetry does support exactly this. Poetry works with pyproject.toml that
was proposed in PEP 518.