Add arg `skip_schema_validation` to @on()
The argument can be used to skip validation of a request and response.
It defaults to `False`, so by default validation is enabled.

The reasoning for this change is that TMH has charge points which use
measurands inside MeterValues messages which are not compliant with the
OCPP 1.6 specification. Therefore the validation of these messages fail.

With this new argument validation of certain messages can be disabled.

Fixes: #54