Log in to Docker hub to prevent being rate limited
Docker hub is changing it's policy as of November 1 2020. It will start
to rate limit pulls of images for non-paying users. This repo relies on
Docker hub for it's automated tests. In order to avoid to be rated
limited this the CircleCI job should log in to Docker hub before pulling
the Python images.