Use IdTagInfo datatype instead of Dict for id_tag_info occurances (#278)
The `Authorize.conf` message in the v1.6 spec defines what fields should
be present for an `IdTagInfo`. The dataclass for this already exists in
`datatypes.py`, as of the recently merged #250 PR.

This PR makes use of it by importing the dataclass, and using it instead
of the more losely typed `typing.Dict`.