Update ChangeConfiguration.json
Dear project owner,
Actually we have found that in the real world case, certain provider, for example Scame from Italy, configure the charger in such a way that the value of configuration could be a string, a number or a boolean. For example, the hearbeat interval or metervalues interval is of a type of integer. That is why we sincerely proposed this change because otherwise the request sent to the charger will trigger some error. 
Thanks
Yong