Revert threads usage; "add executor to validate_payload"
run_in_executor() is in practice just a fancy name for creating
threads. However, this library is based on asyncio; not
threads. Introducing threads into an application has some serious
consequences. For example, it is no longer safe to use os.fork() (or
subprocess with preexec_fn). Thus with threads, this library can be
incompatible with many existing applications, preventing updates.